added docs for joystick
